+++ lib/src/inputs/src/code_editor.dart
@@ -50,6 +50,12 @@ class ThemedCodeEditor extends StatefulWidget {
   /// [customInsers] is a list of Strings to add a shurtcut to insert them in the editor
   final List<String> customInserts;
 
+  /// [onLintTap] is the function that will be called when the user taps on a lint button
+  final Future<List<LintError>> Function(String)? onLintTap;
+
+  /// [onRunTap] is the function that will be called when the user taps on the run button
+  final Future<String?> Function(String)? onRunTap;
+
   const ThemedCodeEditor({
     super.key,
     this.labelText,
@@ -63,9 +69,12 @@ class ThemedCodeEditor extends StatefulWidget {
     this.onSubmitted,
     this.language = LayrzSupportedLanguage.lcl,
     this.constraints,
+    @Deprecated('Now, if you want to append lintErrors, you should use `onLintTap` to perform the lint')
     this.lintErrors = const [],
-    this.i18n,
+    @Deprecated('The i18n will be extracted automatically from the context') this.i18n,
     this.customInserts = const [],
+    this.onLintTap,
+    this.onRunTap,
   });
 
   @override
@@ -73,8 +82,13 @@ class ThemedCodeEditor extends StatefulWidget {
 }
 
 class _ThemedCodeEditorState extends State<ThemedCodeEditor> {
+  LayrzAppLocalizations? get i18n => LayrzAppLocalizations.of(context);
   late CodeController controller;
   String _value = '';
+  bool get _isLoading => _isLinting || _isRunning;
+
+  bool _isRunning = false;
+  bool _isLinting = false;
 
   Mode get _language {
     switch (widget.language) {
@@ -87,7 +101,7 @@ class _ThemedCodeEditorState extends State<ThemedCodeEditor> {
       case LayrzSupportedLanguage.python:
         return python.python;
       default:
-        return Mode();
+        return Mode(refs: {}, disableAutodetect: true);
     }
   }
 
@@ -123,7 +137,8 @@ class _ThemedCodeEditorState extends State<ThemedCodeEditor> {
     }
   }
 
-  Color? get backgroundColor => _theme['root']!.backgroundColor;
+  Color get backgroundColor => _theme['root']!.backgroundColor ?? Colors.black;
+  Color get textColor => validateColor(color: backgroundColor);
 
   @override
   void initState() {
@@ -145,12 +160,33 @@ class _ThemedCodeEditorState extends State<ThemedCodeEditor> {
       _value = widget.value ?? '';
       controller.text = _value;
     }
+
+    if (widget.lintErrors != oldWidget.lintErrors) {
+      controller.analysisResult = AnalysisResult(issues: widget.lintErrors.map(_formatLintError).toList());
+    }
   }
 
   @override
   void dispose() {
     controller.dispose();
     super.dispose();
+  }
+
+  Issue _formatLintError(LintError error) {
+    String message = t(
+      error.code,
+      {
+        'function': error.function,
+        'element': error.element,
+        'required': error.req,
+        'given': error.given,
+      },
+    );
+    return Issue(
+      line: (error.line ?? 1) - 1,
+      message: message,
+      type: IssueType.error,
+    );
   }
 
   @override
@@ -169,19 +205,160 @@ class _ThemedCodeEditorState extends State<ThemedCodeEditor> {
             children: [
               Padding(
                 padding: const EdgeInsets.symmetric(vertical: 10, horizontal: 20),
-                child: widget.label ??
-                    Text(
-                      widget.labelText ?? '',
-                      style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                            color: validateColor(color: backgroundColor ?? Colors.black),
-                            fontFamily: GoogleFonts.jetBrainsMono().fontFamily,
+                child: Row(
+                  children: [
+                    Expanded(
+                      child: widget.label ??
+                          Text(
+                            widget.labelText ?? '',
+                            style: Theme.of(context).textTheme.bodyMedium?.copyWith(
+                                  color: textColor,
+                                  fontFamily: GoogleFonts.jetBrainsMono().fontFamily,
+                                ),
                           ),
                     ),
+                    const SizedBox(width: 10),
+                    if (widget.onLintTap != null) ...[
+                      ThemedTooltip(
+                        message: i18n?.t('actions.lint') ?? 'Lint',
+                        color: Colors.white,
+                        child: InkWell(
+                          onTap: _isLoading
+                              ? null
+                              : () async {
+                                  setState(() => _isLinting = true);
+                                  controller.analysisResult = const AnalysisResult(issues: []);
+                                  await Future.delayed(const Duration(milliseconds: 500));
+                                  final result = await widget.onLintTap?.call(_value);
+                                  setState(() => _isLinting = false);
+
+                                  if (result != null) {
+                                    controller.analysisResult = AnalysisResult(
+                                      issues: result.map(_formatLintError).toList(),
+                                    );
+                                  }
+                                },
+                          child: SizedBox(
+                            height: 20,
+                            width: 20,
+                            child: Stack(
+                              children: [
+                                Center(
+                                  child: AnimatedOpacity(
+                                    opacity: _isLinting ? 0.5 : 1,
+                                    duration: kHoverDuration,
+                                    child: Icon(
+                                      MdiIcons.bugOutline,
+                                      color: textColor,
+                                      size: 16,
+                                    ),
+                                  ),
+                                ),
+                                Center(
+                                  child: AnimatedOpacity(
+                                    opacity: _isLinting ? 1 : 0,
+                                    duration: kHoverDuration,
+                                    child: CircularProgressIndicator(
+                                      color: textColor,
+                                      strokeWidth: 2,
+                                    ),
+                                  ),
+                                ),
+                              ],
+                            ),
+                          ),
+                        ),
+                      ),
+                      const SizedBox(width: 10),
+                    ],
+                    if (widget.onRunTap != null)
+                      ThemedTooltip(
+                        message: i18n?.t('actions.run') ?? 'Run',
+                        color: Colors.white,
+                        child: InkWell(
+                          onTap: _isLoading
+                              ? null
+                              : () async {
+                                  setState(() => _isRunning = true);
+                                  await Future.delayed(const Duration(milliseconds: 500));
+                                  final result = await widget.onRunTap?.call(_value);
+                                  setState(() => _isRunning = false);
+
+                                  if (result != null) {
+                                    showDialog(
+                                      context: context,
+                                      builder: (context) {
+                                        return Dialog(
+                                          elevation: 0,
+                                          backgroundColor: Colors.transparent,
+                                          child: Container(
+                                            decoration: generateContainerElevation(
+                                              context: context,
+                                              color: backgroundColor,
+                                              elevation: 3,
+                                            ),
+                                            constraints: const BoxConstraints(maxWidth: 400),
+                                            padding: const EdgeInsets.all(20),
+                                            child: Text(
+                                              result,
+                                              style: Theme.of(context).textTheme.bodyMedium?.copyWith(
+                                                    color: textColor,
+                                                    fontFamily: GoogleFonts.jetBrainsMono().fontFamily,
+                                                  ),
+                                              maxLines: 4,
+                                              textAlign: TextAlign.center,
+                                            ),
+                                          ),
+                                        );
+                                      },
+                                    );
+                                  }
+                                },
+                          child: SizedBox(
+                            height: 20,
+                            width: 20,
+                            child: Stack(
+                              children: [
+                                Center(
+                                  child: AnimatedOpacity(
+                                    opacity: _isRunning ? 0.5 : 1,
+                                    duration: kHoverDuration,
+                                    child: Padding(
+                                      padding: const EdgeInsets.only(left: 2, bottom: 2),
+                                      child: Icon(
+                                        MdiIcons.playOutline,
+                                        color: textColor,
+                                        size: 18,
+                                      ),
+                                    ),
+                                  ),
+                                ),
+                                Center(
+                                  child: AnimatedOpacity(
+                                    opacity: _isRunning ? 1 : 0,
+                                    duration: kHoverDuration,
+                                    child: CircularProgressIndicator(
+                                      color: textColor,
+                                      strokeWidth: 2,
+                                    ),
+                                  ),
+                                ),
+                              ],
+                            ),
+                          ),
+                        ),
+                      ),
+                  ],
+                ),
               ),
-              Divider(
-                color: validateColor(color: backgroundColor ?? Colors.black).withOpacity(0.2),
-                indent: 10,
-                endIndent: 10,
+              Padding(
+                padding: const EdgeInsets.symmetric(horizontal: 20),
+                child: LinearProgressIndicator(
+                  value: _isLoading ? null : 0,
+                  minHeight: 1,
+                  backgroundColor: textColor.withOpacity(0.2),
+                  color: textColor,
+                ),
               ),
               Theme(
                 data: ThemeData.dark().copyWith(
@@ -201,19 +378,7 @@ class _ThemedCodeEditorState extends State<ThemedCodeEditor> {
               ),
               ThemedFieldDisplayError(
                 padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 10),
-                errors: widget.errors +
-                    widget.lintErrors.map<String>((error) {
-                      String err = t(
-                        'larzLanguage.errors.${error.code}',
-                        {
-                          'function': error.function,
-                          'element': error.element,
-                          'required': error.req,
-                          'given': error.given,
-                        },
-                      );
-                      return "${error.line ?? 1}: $err";
-                    }).toList(),
+                errors: widget.errors,
               ),
             ],
           ),
@@ -223,8 +388,41 @@ class _ThemedCodeEditorState extends State<ThemedCodeEditor> {
   }
 
   String t(String key, [Map<String, dynamic> args = const {}]) {
-    return widget.i18n?.t(key, args) ?? "I18n missing $key";
+    if (i18n != null) return i18n!.t(key, args);
+
+    if (_defaultErrors.containsKey(key)) {
+      String message = _defaultErrors[key]!;
+      args.forEach((key, value) {
+        message = message.replaceAll('{$key}', value.toString());
+      });
+
+      return message;
+    }
+
+    return "Translation missing $key";
   }
+
+  Map<String, String> get _defaultErrors => {
+        'linter.parse.syntax.error': 'Syntax error',
+
+        // LCL errors
+        'lcl.function.arguments.mismatch': 'Function {function} arguments mismatch, required {required} given {given}',
+        'lcl.function.not.found': 'Function {function} not found',
+
+        // Python errors
+        'python.main.function.not.found': 'Main function not found',
+        'python.main.function.arguments.mismatch': 'Main function arguments mismatch',
+        'python.return.not.allowed.outside.function': 'Return not allowed outside function',
+        'python.prohibited.element.used': 'Prohibited element {element} used',
+        'python.async.function.not.allowed': 'Asyncronus functions are not allowed',
+        'python.yield.not.allowed': 'Yield is not allowed',
+        'python.infinite.value.not.allowed': 'Infinite value is not allowed',
+        'python.print.not.allowed': 'Print statement is not allowed',
+        'python.plpy.not.allowed': 'Plpy module is not allowed',
+        'python.import.not.allowed': 'Import statement is not allowed',
+        'python.while.not.allowed': 'While statement is not allowed',
+        'python.raise.not.allowed': 'Raise statement is not allowed',
+      };
 }
 
 enum LayrzSupportedLanguage {