+++ lib/src/languages/lcl/src/analyzer.dart
@@ -3,7 +3,6 @@ part of '../lcl.dart';
 class LclAnalizer extends AbstractAnalyzer {
   @override
   Future<AnalysisResult> analyze(Code code) async {
-    // TODO Develop an analyzer for LCL
     return const AnalysisResult(issues: []);
   }
 }