+++ lib/src/languages/lml/src/analyzer.dart
@@ -3,7 +3,6 @@ part of '../lml.dart';
 class LmlAnalizer extends AbstractAnalyzer {
   @override
   Future<AnalysisResult> analyze(Code code) async {
-    // TODO Develop an analyzer for LML
     return const AnalysisResult(issues: []);
   }
 }