+++ lib/src/languages/mjml/src/analyzer.dart
@@ -3,7 +3,6 @@ part of '../mjml.dart';
 class MjmlAnalizer extends AbstractAnalyzer {
   @override
   Future<AnalysisResult> analyze(Code code) async {
-    // TODO Develop an analyzer for MJML
     return const AnalysisResult(issues: []);
   }
 }