+++ lib/src/languages/python/src/analyzer.dart
@@ -3,7 +3,6 @@ part of '../python.dart';
 class PythonAnalizer extends AbstractAnalyzer {
   @override
   Future<AnalysisResult> analyze(Code code) async {
-    // TODO Develop an analyzer for Python
     return const AnalysisResult(issues: []);
   }
 }